Broke down some code into components, Update now uses an API from azure instead of a local url.
